test attaching process

+++ src/process.cpp
@@ -25,16 +25,17 @@ void exit_with_perror(gsdb::pipe& channel, std::string const& prefix) {
 gsdb::process::~process() {
     if (pid_ != 0) {
         int status;
-        if (state_ == process_state::running) {
-            kill(pid_, SIGSTOP);
-            waitpid(pid_, &status, 0);
+        if (is_attached_) {
+            if (state_ == process_state::running) {
+                kill(pid_, SIGSTOP);
+                waitpid(pid_, &status, 0);
+            }
+
+            // for PTRACE_DETACH to work, the inferior must be stopped
+            // thus the SIGSTOP above
+            ptrace(PTRACE_DETACH, pid_, nullptr, nullptr);
+            kill(pid_, SIGCONT);  // detached, then let it continue
         }
-
-        // for PTRACE_DETACH to work, the inferior must be stopped
-        // thus the SIGSTOP above
-        ptrace(PTRACE_DETACH, pid_, nullptr, nullptr);
-        kill(pid_, SIGCONT);  // detached, then let it continue
-
         if (terminate_on_end_) {
             kill(pid_, SIGKILL);
             waitpid(pid_, &status, 0);  // wait for it to terminate
@@ -76,8 +77,8 @@ gsdb::stop_reason gsdb::process::wait_on_signal() {
     return reason;
 }
 
-std::unique_ptr<gsdb::process> gsdb::process::launch(
-    std::filesystem::path path) {
+std::unique_ptr<gsdb::process> gsdb::process::launch(std::filesystem::path path,
+                                                     bool debug) {
     // IMPORTANT: CALL pipe BEFORE fork
     pipe channel(/*close_on_exec=*/true);
 
@@ -89,8 +90,7 @@ std::unique_ptr<gsdb::process> gsdb::process::launch(
     if (pid == 0) {
         // child process
         channel.close_read();
-        if (ptrace(PTRACE_TRACEME, 0, nullptr, nullptr) < 0) {
-            error::send_errno("Tracing failed");
+        if (debug and ptrace(PTRACE_TRACEME, 0, nullptr, nullptr) < 0) {
             exit_with_perror(channel, "Tracing failed");
         }
         if (execlp(path.c_str(), path.c_str(), nullptr) < 0) {
@@ -111,8 +111,14 @@ std::unique_ptr<gsdb::process> gsdb::process::launch(
         error::send(std::string(chars, chars + data.size()));
     }
 
-    std::unique_ptr<process> proc(new process(pid, /*terminate_on_end=*/true));
-    proc->wait_on_signal();  // wait for the process to halt
+    std::unique_ptr<process> proc(
+        new process(pid, /*terminate_on_end=*/true, debug));
+    // if we don't start tracing we should _not_ wait for a signal after
+    // forking, because the process won't stop automatically after the `exec`
+    // call
+    if (debug) {
+        proc->wait_on_signal();  // wait for the process to halt
+    }
 
     return proc;
 }
@@ -126,7 +132,8 @@ std::unique_ptr<gsdb::process> gsdb::process::attach(pid_t pid) {
         error::send_errno("Could NOT attach");
     }
 
-    std::unique_ptr<process> proc(new process(pid, /*terminate_on_end=*/false));
+    std::unique_ptr<process> proc(
+        new process(pid, /*terminate_on_end=*/false, /*is_attached=*/true));
     proc->wait_on_signal();  // wait for the underlying process to halt
 
     return proc;
